requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==15.0.0
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
ctranslate2==4.6.0
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
fastapi==0.116.0
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
identify==2.6.12
idna==3.10
jiter==0.10.0
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8.2
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
onnxruntime==1.20.1
openai==1.70.0
packaging==25.0
pathspec==0.12.1
pillow==11.1.0
pipecat-ai==0.0.74
platformdirs==4.3.8
pre_commit==4.2.0
propcache==0.3.2
protobuf==5.29.5
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyloudnorm==0.1.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
resampy==0.4.3
scipy==1.16.0
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.46.2
structlog==25.4.0
sympy==1.14.0
tokenizers==0.21.2
tqdm==4.67.1
twilio==9.6.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
